Create README.md
Graph Theory is used in various sciences, computer science also tends to solve a lot of problems with
Graph Theory. Some of the applications of Graph Theory include social media problems, travel, chip design,
and many other fields.
Social media app, such as LinkedIn, which needs to connect millions of people eventually it will create
enormous number of edges. To be able to apply analytics on this kind of data, regular databases will not
suffice. With a regular database, we would need to apply self-joins so many times and applying self-joins will
literally bring our database down so the optimal solution is graph theory application.